requirements.txt
flask
Werkzeug
flask_compress
flask_wtf
flask_sslify
python-docx
fpdf
pillow
pytesseract